During importance sampling ensure update blocks are placed inside the problem area if they would overlap the edge
Prior to this change the origin of the update block could be placed at the outer edge of a problem, for example if the
neumann boundary is at z==0. This would place the rest of the block outside the problem area. With this change the block origin
will be offset into the problem so that the entire update area is inside the problem space.