Reimplemented residuals and importance sampling
Currently residuals are not "true" residuals but more a rough estimate of where the next update blocks
should be positioned. A periodic true residuals pass needs to be implemented next to calculate the real
residuals of *all* vertices once every n iterations, to check for convergence and correct inaccuracies in
the 'fake' residuals.